ingestion/drive_service/requirements.txt
google-api-python-client==2.139.0
google-auth==2.32.0
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.18.2
google-cloud-logging==3.11.1